cli: move the shared flash machinery out of the Thor file
The step UI, download verification and recovery wait were already shared with
Orin; a third board made the Thor-named home misleading. Their Jetson wording
is now a hint field so a non-Jetson target can supply its own.
